+++ src/lib.rs
@@ -1,9 +1,13 @@
+use std::collections::HashMap;
 use std::ffi::OsStr;
+use std::io::BufRead;
+use std::sync::atomic::AtomicBool;
 use std::{fs::File, process::Command};
 
 use anyhow::{anyhow, bail, Context as _, Error, Result};
 use camino::{Utf8Path, Utf8PathBuf};
 use clap::{Arg, ArgAction, ArgMatches, ValueHint};
+use rayon::{prelude::*, ThreadPoolBuilder};
 
 pub struct GitCache {
     cache_base_dir: Utf8PathBuf,
@@ -41,9 +45,17 @@ pub struct GitCacheCloner {
     #[builder(default)]
     sparse_paths: Option<Vec<String>>,
     #[builder(default)]
+    recurse_submodules: Option<Vec<String>>,
+    #[builder(default)]
+    recurse_all_submodules: bool,
+    #[builder(default)]
+    shallow_submodules: bool,
+    #[builder(default)]
     commit: Option<String>,
     #[builder(default)]
     extra_clone_args: Option<Vec<String>>,
+    #[builder(default)]
+    jobs: Option<usize>,
 }
 
 impl GitCacheClonerBuilder {
@@ -140,20 +152,58 @@ impl GitCacheCloner {
             )?;
         }
 
+        let target_repo = GitRepo {
+            path: target_path.clone(),
+        };
+
         if let Some(commit) = wanted_commit {
-            let target_repo = GitRepo {
-                path: target_path.clone(),
-            };
             target_repo.set_config("advice.detachedHead", "false")?;
             target_repo.checkout(commit)?;
         }
         if let Some(sparse_paths) = self.sparse_paths.as_ref() {
-            let target_repo = GitRepo {
-                path: target_path.clone(),
-            };
             target_repo.sparse_checkout(sparse_paths)?;
         }
+
+        if self.recurse_all_submodules || self.recurse_submodules.is_some() {
+            let filter = if !self.recurse_all_submodules {
+                self.recurse_submodules.clone()
+            } else {
+                None
+            };
+
+            let cache = self.cache()?;
+
+            let jobs = self.jobs.unwrap_or(1);
+
+            static RAYON_CONFIGURED: AtomicBool = AtomicBool::new(false);
+
+            if !RAYON_CONFIGURED.swap(true, std::sync::atomic::Ordering::AcqRel) {
+                let _ = ThreadPoolBuilder::new().num_threads(jobs).build_global();
+            }
+
+            target_repo
+                .get_submodules(filter)?
+                .par_iter()
+                .map(|submodule| {
+                    println!(
+                        "git-cache: cloning {} into {}...",
+                        submodule.url, submodule.path
+                    );
+                    target_repo.clone_submodule(
+                        submodule,
+                        &cache,
+                        self.shallow_submodules,
+                        self.update,
+                    )
+                })
+                .collect::<Result<Vec<_>, _>>()?;
+        };
+
         Ok(())
+    }
+
+    pub fn cache(&self) -> Result<GitCache, anyhow::Error> {
+        GitCache::new(self.cache_base_dir.clone())
     }
 }
 
@@ -216,6 +266,23 @@ impl GitRepo {
             .true_or(anyhow!("error checking out commit"))
     }
 
+    fn submodule_commits(&self) -> Result<HashMap<String, String>> {
+        let output = self.git().arg("submodule").arg("status").output()?;
+
+        let res = output
+            .stdout
+            .lines()
+            .map(|line| line.unwrap())
+            .map(|line| {
+                // ` f47ce7b5fbbb3aa43d33d2be1f6cd3746b13d5bf some/path`
+                let commit = line[1..41].to_string();
+                let path = line[42..].to_string();
+                (path, commit)
+            })
+            .collect::<HashMap<String, String>>();
+        Ok(res)
+    }
+
     fn sparse_checkout<I, S>(&self, sparse_paths: I) -> std::result::Result<(), anyhow::Error>
     where
         I: IntoIterator<Item = S>,
@@ -228,6 +295,106 @@ impl GitRepo {
             .status()?
             .success()
             .true_or(anyhow!("error setting up sparse checkout"))
+    }
+
+    fn get_submodules(
+        &self,
+        filter: Option<Vec<String>>,
+    ) -> std::result::Result<Vec<SubmoduleSpec>, anyhow::Error> {
+        use gix_config::File;
+        let mut path = self.path.clone();
+        path.push(".gitmodules");
+
+        if !path.exists() {
+            return Ok(Vec::new());
+        }
+
+        let gitconfig = File::from_path_no_includes(path.into(), gix_config::Source::Api)?;
+        let gitmodules = gitconfig.sections_by_name("submodule");
+
+        if gitmodules.is_none() {
+            return Ok(Vec::new());
+        }
+
+        let submodule_commits = self.submodule_commits()?;
+
+        println!("{:?}", submodule_commits);
+
+        let mut submodules = Vec::new();
+        for module in gitmodules.unwrap() {
+            let path = module.body().value("path");
+            let url = module.body().value("url");
+            let branch = module.body().value("branch").map(|b| b.to_string());
+
+            if path.is_none() || url.is_none() {
+                eprintln!("git-cache: submodule missing path or url");
+                continue;
+            }
+            let path = path.unwrap().into_owned().to_string();
+            let url = url.unwrap().into_owned().to_string();
+
+            let commit = submodule_commits.get(&path);
+
+            if commit.is_none() {
+                eprintln!("git-cache: could not find submodule commit for path `{path}`");
+            }
+
+            if let Some(filter) = filter.as_ref() {
+                if !filter.contains(&path) {
+                    continue;
+                }
+            }
+
+            submodules.push(SubmoduleSpec::new(
+                path,
+                url,
+                commit.unwrap().clone(),
+                branch,
+            ));
+        }
+
+        Ok(submodules)
+    }
+
+    fn clone_submodule(
+        &self,
+        submodule: &SubmoduleSpec,
+        cache: &GitCache,
+        shallow_submodules: bool,
+        update: bool,
+    ) -> std::result::Result<(), anyhow::Error> {
+        let submodule_path = self.path.join(&submodule.path);
+
+        let mut cloner = cache.cloner();
+
+        cloner
+            .repository_url(submodule.url.clone())
+            .target_path(Some(submodule_path))
+            .recurse_all_submodules(true)
+            .shallow_submodules(shallow_submodules)
+            .commit(Some(submodule.commit.clone()))
+            .update(update);
+
+        // if let Some(branch) = submodule.branch {
+        //     cloner.extra_clone_args(Some(vec!["--branch".into(), branch]));
+        // }
+
+        cloner.do_clone()?;
+
+        self.init_submodule(&submodule.path)?;
+
+        Ok(())
+    }
+
+    fn init_submodule(&self, path: &str) -> std::result::Result<(), anyhow::Error> {
+        self.git()
+            .arg("submodule")
+            .arg("init")
+            .arg("--")
+            .arg(path)
+            .status()?
+            .success()
+            .true_or(anyhow!("error initializing submodule"))
     }
 }
 
@@ -408,6 +575,37 @@ pub fn clap_clone_command(name: &'static str) -> clap::Command {
                 .action(ArgAction::Append)
                 .help("do a sparse checkout, keep PATH"),
         )
+        .arg(
+            Arg::new("recurse-submodules")
+                .long("recurse-submodules")
+                .value_name("pathspec")
+                .action(ArgAction::Append)
+                .num_args(0..=1)
+                .require_equals(true)
+                .help("recursively clone submodules"),
+        )
+        .arg(
+            Arg::new("shallow-submodules")
+                .long("shallow-submodules")
+                .action(ArgAction::SetTrue)
+                .overrides_with("no-shallow-submodules")
+                .help("shallow-clone submodules"),
+        )
+        .arg(
+            Arg::new("no-shallow-submodules")
+                .long("no-shallow-submodules")
+                .action(ArgAction::SetTrue)
+                .overrides_with("shallow-submodules")
+                .help("don't shallow-clone submodules"),
+        )
+        .arg(
+            Arg::new("jobs")
+                .long("jobs")
+                .short('j')
+                .help("The number of submodules fetched at the same time.")
+                .num_args(1)
+                .value_parser(clap::value_parser!(usize)),
+        )
         .args(pass_through_args())
         .after_help(
             "These regular \"git clone\" options are passed through:\n
@@ -466,7 +664,6 @@ fn pass_through_args() -> Vec<Arg> {
     for (short, long) in [
         ('b', "branch"),
         ('c', "config"),
-        ('j', "jobs"),
         ('o', "origin"),
         ('u', "upload-pack"),
     ]
@@ -490,12 +687,10 @@ fn pass_through_args() -> Vec<Arg> {
         "no-hardlinks",
         "no-reject-shallow",
         "no-remote-submodules",
-        "no-shallow-submodules",
         "no-single-branch",
         "no-tags",
         "reject-shallow",
         "remote-submodules",
-        "shallow-submodules",
         "single-branch",
     ]
     .into_iter()
@@ -520,15 +715,6 @@ fn pass_through_args() -> Vec<Arg> {
         args.push(Arg::new(id).long(id).num_args(1).hide(true));
     }
 
-    // long with optional arg
-    args.push(
-        Arg::new("recurse-submodules")
-            .long("recurse-submodules")
-            .num_args(0..=1)
-            .action(ArgAction::Append)
-            .hide(true),
-    );
-
     args
 }
 
@@ -548,12 +734,10 @@ fn get_pass_through_args(matches: &ArgMatches) -> Vec<String> {
         "no-hardlinks",
         "no-reject-shallow",
         "no-remote-submodules",
-        "no-shallow-submodules",
         "no-single-branch",
         "no-tags",
         "reject-shallow",
         "remote-submodules",
-        "shallow-submodules",
         "single-branch",
         "sparse",
     ]
@@ -571,7 +755,6 @@ fn get_pass_through_args(matches: &ArgMatches) -> Vec<String> {
         "config",
         "depth",
         "filter",
-        "jobs",
         "origin",
         "reference",
         "reference-if-able",
@@ -587,16 +770,6 @@ fn get_pass_through_args(matches: &ArgMatches) -> Vec<String> {
             for occurrence in occurrences.flatten() {
                 args.push(format!("--{id}"));
                 args.push(occurrence.clone());
-            }
-        }
-    }
-
-    if let Some(submodules) = matches.get_many::<String>("recurse-submodules") {
-        if submodules.len() == 0 && matches.contains_id("recurse-submodules") {
-            args.push("--recurse-submodules".into());
-        } else {
-            for submodule in submodules {
-                args.push(format!("--recurse-submodules={submodule}"));
             }
         }
     }
@@ -624,6 +797,26 @@ impl TrueOr for bool {
             Ok(())
         } else {
             Err(error)
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+struct SubmoduleSpec {
+    path: String,
+    url: String,
+    #[allow(dead_code)]
+    branch: Option<String>,
+    commit: String,
+}
+
+impl SubmoduleSpec {
+    pub fn new(path: String, url: String, commit: String, branch: Option<String>) -> Self {
+        Self {
+            path,
+            url,
+            commit,
+            branch,
         }
     }
 }