+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+- implement recursive submodule cloning from cache
+
 ## [0.2.3] - 2024-07-01
 
 ## [0.2.2] - 2024-07-01